Fix JSON number type detection bug and update tests for batch processing refactoring
- Fix inverted Integer/Double type detection in JRPC.Core
- Fix exception propagation in JRPC.Invoker: re-raise EJRPCException as-is
- Update tests to use new type hierarchy

+++ Source/MCPConnect.JRPC.Core.pas
@@ -838,9 +838,9 @@ begin
   if AValue is TJSONNumber then
   begin
     if (AValue as TJSONNumber).Value.Contains('.') then
-      LType := TRttiUtils.Context.GetType(TypeInfo(Integer))
+      LType := TRttiUtils.Context.GetType(TypeInfo(Double))
     else
-      LType := TRttiUtils.Context.GetType(TypeInfo(Double));
+      LType := TRttiUtils.Context.GetType(TypeInfo(Integer));
 
     LValue := AContext.ReadDataMember(AValue, LType, AData, False);
   end
